+++ js/constants.js
@@ -1,0 +1,46 @@
+const ADVERTISEMENT_COUNT = 10;
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const CHECKINS = ['12:00', '13:00', '14:00'];
+const CHECKOUTS = ['12:00', '13:00', '14:00'];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const MIN = 1;
+const MAX = 1000;
+const LATMIN = 35.65;
+const LATMAX = 35.7;
+const LAT_LNG_PRECISION = 5;
+const LNGMIN = 139.7;
+const LNGMAX = 139.8;
+const MINLENGTH = 1;
+const AVATARMININDEX = 1;
+const AVATARMAXINDEX = 10;
+
+export {
+  ADVERTISEMENT_COUNT,
+  TYPES,
+  CHECKINS,
+  CHECKOUTS,
+  FEATURES,
+  PHOTOS,
+  MIN,
+  MAX,
+  LATMIN,
+  LATMAX,
+  LAT_LNG_PRECISION,
+  LNGMIN,
+  LNGMAX,
+  MINLENGTH,
+  AVATARMININDEX,
+  AVATARMAXINDEX
+};
