--- js/moduls/activate-page.js
@@ -1,3 +1,0 @@
-const activatePage = () => {};
-
-export {activatePage};
