--- js/moduls/deactivate-page.js
@@ -1,3 +1,0 @@
-const deactivatePage = () => {};
-
-export {deactivatePage};
