--- js/moduls/fill-form-post-data.js
@@ -1,4 +1,0 @@
-const fillForm = () => {};
-const postData = () => {};
-
-export {fillForm, postData};
