--- js/moduls/map.js
@@ -1,4 +1,0 @@
-const isMapLoaded = () => {};
-const returnSearchAddressCoordinates = () => {};
-
-export {isMapLoaded, returnSearchAddressCoordinates};
